requirements.txt
openpyxl==3.1.2
pandas==1.5.2
streamlit.egg==info
streamlit_authenticator==0.1.5
streamlit_option_menu==0.3.12
